* dev/misc-cleanups:
  Return TRUE from accepted_languages in WebKit2
  Make webkit_view available in WebKit2
  Don't app_cache_directory in WebKit2
  Always free the path memory
  Free I/O resources
  Set the io_loop to NULL after destroying
  Minimize the number of paths for returning
  Add missing brace
  Cast the decision in WebKit2
  Avoid warnings with WebKit2 in `snapshot`
  Use doubles instead of floats

+++ uzbl/plugins/config.py
@@ -3,7 +3,7 @@ from re import compile
 from uzbl.arguments import splitquoted
 from uzbl.ext import PerInstancePlugin
 
-types = {'int': int, 'float': float, 'str': str, 'ull': int}
+types = {'int': int, 'double': float, 'str': str, 'ull': int}
 
 valid_key = compile('^[A-Za-z0-9_\.]+$').match
